amount fix to minor currency unit

+++ src/main/java/com/checkout/payment/gateway/model/PostPaymentResponse.java
@@ -4,7 +4,6 @@ import com.checkout.payment.gateway.enums.PaymentStatus;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import java.util.UUID;
-import java.math.BigDecimal;
 
 @Data
 public class PostPaymentResponse {
@@ -20,6 +19,6 @@ public class PostPaymentResponse {
     private int expiryYear;
     @JsonProperty("currency")              
     private String currency;
-    @JsonProperty("amount")                
-    private BigDecimal amount;
+    @JsonProperty("amount")
+    private int amount;
 }
